simplified modal.js and removed close-modal.js
