+++ index.html
@@ -102,7 +102,6 @@
   <script type="module" src="/js/render/home-page/movies.js"></script>
   <script type="module" src="/js/render/home-page/movie-list-page.js"></script>
   <script type="module" src="/js/render/shared/modal.js"></script>
-  <script type="module" src="/js/render/shared/close-modal.js"></script>
   <script type="module" src="/js/init/index.js"></script>
 </body> 
 </html>
