+++ js/render/shared/modal.js
@@ -1,6 +1,6 @@
-import { listMovie } from "../../api/api.js";
-import { modalComponent } from "../../components/_modal.js";
-import { renderError } from "../../render/shared/error.js";
+import {listMovie} from "../../api/api.js";
+import {modalComponent} from "../../components/_modal.js";
+import {renderError} from "../../render/shared/error.js";
 
 const $container = document.querySelector("#container");
 const $relatedMoviesContainer = document.querySelector("#related-movies");
@@ -16,7 +16,9 @@ async function renderModalContent(imdbID) {
 }
 
 if ($container) $container.addEventListener("click", openModal);
+if ($container) $container.addEventListener("click", closeModal);
 if ($relatedMoviesContainer) $relatedMoviesContainer.addEventListener("click", openModal);
+if ($relatedMoviesContainer) $relatedMoviesContainer.addEventListener("click", closeModal);
 
 function openModal(event) {
   const openModalButton = event.target.closest("[data-modal-btn-open-id]");
@@ -27,4 +29,21 @@ function openModal(event) {
   }
 }
 
-export { renderModalContent };
+function clearModalContent(imdbID) {
+  const $modal = document.querySelector(`[data-modal-open-id=${imdbID}]`);
+  const fadeOutMs = parseInt(getComputedStyle(document.documentElement).getPropertyValue('--fade-ms'));
+  $modal.classList.add("hide-modal");
+
+  setTimeout(() => {
+    $modal.remove();
+  }, fadeOutMs);
+}
+
+function closeModal(event) {
+  const closeModalButton = event.target.closest("[data-modal-btn-close-id]");
+  
+  if (closeModalButton) {
+    const imdbID = closeModalButton.dataset.modalBtnCloseId;
+    clearModalContent(imdbID);
+  }
+}
