+++ movie.html
@@ -124,7 +124,6 @@
   <script type="module" src="/js/render/movie-details-page/related-movies.js"></script>
   <script type="module" src="/js/render/movie-details-page/movie-page.js"></script>
   <script type="module" src="/js/render/shared/modal.js"></script>
-  <script type="module" src="/js/render/shared/close-modal.js"></script>
   <script type="module" src="/js/init/movie.js"></script>
 </body>
