+++ scss/components/_card.scss
@@ -14,7 +14,7 @@
 
   img {
     width: 100%;
-    min-height: 48rem;
+    min-height: 49rem;
     max-height: 49rem;
     display: block;
     object-fit: cover;
